requirements.txt
fastapi==0.115.13
uvicorn==0.32.1
asyncpg==0.30.0
python-dotenv==1.1.0
pydantic-settings==2.10.1
pydantic==2.10.3
